move strings and refactor mos_exec
common code to run a binary at an address made into a function
