+++ src/strings.c
@@ -6,6 +6,8 @@
  */
  
 #include <ctype.h>
+#include <stdlib.h>
+#include <string.h>
 
 int strcasecmp(const char *s1, const char *s2)
 {
@@ -22,3 +24,31 @@ int strcasecmp(const char *s1, const char *s2)
 	return result;
 }
 
+//Alternative to missing strnlen() in ZDS libraries
+size_t mos_strnlen(const char *s, size_t maxlen) {
+    size_t len = 0;
+    while (len < maxlen && s[len] != '\0') {
+        len++;
+    }
+    return len;
+}
+
+//Alternative to missing strdup() in ZDS libraries
+char *mos_strdup(const char *s) {
+    char *d = malloc(strlen(s) + 1);  // Allocate memory
+    if (d != NULL) strcpy(d, s);      // Copy the string
+    return d;
+}
+
+//Alternative to missing strndup() in ZDS libraries
+char *mos_strndup(const char *s, size_t n) {
+    size_t len = mos_strnlen(s, n);
+    char *d = malloc(len + 1);  // Allocate memory for length plus null terminator
+
+    if (d != NULL) {
+        strncpy(d, s, len);  // Copy up to len characters
+        d[len] = '\0';       // Null-terminate the string
+    }
+
+    return d;
+}
