+++ src/strings.h
@@ -12,4 +12,12 @@
 
 extern int strcasecmp(const char *s1, const char *s2);
 
+size_t mos_strnlen(const char *s, size_t maxlen);
+
+//Alternative to missing strdup() in ZDS libraries
+char *mos_strdup(const char *s);
+
+//Alternative to missing strndup() in ZDS libraries
+char *mos_strndup(const char *s, size_t n);
+
 #endif // STRINGS_H
